Penambahan route middleware untuk hak akses setiap menu pada backend

+++ sco-server/routes/api.php
@@ -32,60 +32,69 @@ Route::group(['middleware' => ['auth:sanctum']], function () {
      * Route group untuk menu item
      */
     Route::group(['prefix' => 'menu'], function () {
-        Route::get('/', [MenuItemController::class, 'index']);
-        Route::post('/', [MenuItemController::class, 'store']);
-        Route::put('/{menuItem}', [MenuItemController::class, 'update']);
-        Route::delete('/{menuItem}', [MenuItemController::class, 'destroy']);
+        Route::get('/', [MenuItemController::class, 'index'])->middleware('access.menu:read');
+        Route::post('/', [MenuItemController::class, 'store'])->middleware('access.menu:create');
+        Route::put('/{menuItem}', [MenuItemController::class, 'update'])->middleware('access.menu:update');
+        Route::delete('/{menuItem}', [MenuItemController::class, 'destroy'])->middleware('access.menu:delete');
     });
 
     /**
-     * Route group untuk menu item
+     * Route group untuk menu sub item
      */
     Route::group(['prefix' => 'submenu'], function () {
-        Route::get('/', [MenuSubItemController::class, 'index']);
-        Route::post('/', [MenuSubItemController::class, 'store']);
-        Route::put('/{menuSubItem}', [MenuSubItemController::class, 'update']);
-        Route::delete('/{menuSubItem}', [MenuSubItemController::class, 'destroy']);
+        Route::get('/', [MenuSubItemController::class, 'index'])->middleware('access.menu:read');
+        Route::post('/', [MenuSubItemController::class, 'store'])->middleware('access.menu:create');
+        Route::put('/{menuSubItem}', [MenuSubItemController::class, 'update'])->middleware('access.menu:update');
+        Route::delete('/{menuSubItem}', [MenuSubItemController::class, 'destroy'])->middleware('access.menu:delete');
     });
 
     /**
      * Route group untuk user
      */
     Route::group(['prefix' => 'user'], function () {
-        Route::get('/', [UserController::class, 'index']);
-        Route::get('/menu', [UserController::class, 'getMenus']);
-        Route::get('/{user}', [UserController::class, 'show']);
-        Route::get('/{id}/edit', [UserController::class, 'editProfile']);
-        Route::get('/{user}/account', [UserController::class, 'editAccount']);
-
-        Route::post('/cek/user-form', [UserController::class, 'cekUserForm']);
-        Route::post('/cek/profile-form', [UserController::class, 'cekProfileForm']);
-        Route::post('/', [UserController::class, 'store']);
-        Route::post('/menu-access', [UserController::class, 'createUserMenuAccess']);
-
-        Route::delete('/truncate-tokens', [UserController::class, 'truncateTokens']);
-        Route::delete('/{id}', [UserController::class, 'destroy']);
-
-        Route::patch('/{id}', [UserController::class, 'updateProfile']);
-        Route::patch('/{user}/account', [UserController::class, 'updateAccount']);
 
         /**
-         * Route edit user menu item akses
+         * Route middleware menu user untuk akses read
          */
-        Route::get('/menu/{id}', [UserController::class, 'getUserMenuItems']);
-        Route::post('/menu/{id}', [UserController::class, 'addUserMenuItem']);
-        Route::delete('/menu/{id}', [UserController::class, 'deleteUserMenuItem']);
+        Route::group(['middleware' => ['access.user:read']], function () {
+            Route::get('/', [UserController::class, 'index']);
+            Route::get('/menu', [UserController::class, 'getMenus']);
+            Route::get('/{user}', [UserController::class, 'show']);
+        });
 
         /**
-         * Route edit user menu sub item akses
+         * Route middleware menu user untuk akses create
          */
-        Route::get('/submenu/{id}', [UserController::class, 'getUserMenuSubItems']);
-        Route::post('/submenu/{id}', [UserController::class, 'addUserMenuSubItems']);
-        Route::delete('/submenu/{id}', [UserController::class, 'deleteUserMenuSubItems']);
+        Route::group(['middleware' => ['access.user:create']], function () {
+            Route::post('/cek/user-form', [UserController::class, 'cekUserForm']);
+            Route::post('/cek/profile-form', [UserController::class, 'cekProfileForm']);
+            Route::post('/', [UserController::class, 'store']);
+            Route::post('/menu-access', [UserController::class, 'createUserMenuAccess']);
+        });
 
         /**
-         * Route untuk merubah password user
+         * Route middleware menu user untuk akses update
          */
-        Route::patch('/password/{id}', [UserController::class, 'updatePassword']);
+        Route::group(['middleware' => ['access.user:update']], function () {
+            Route::get('/{id}/edit', [UserController::class, 'editProfile']);
+            Route::get('/{user}/account', [UserController::class, 'editAccount']);
+            Route::patch('/{id}', [UserController::class, 'updateProfile']);
+            Route::patch('/{user}/account', [UserController::class, 'updateAccount']);
+            Route::get('/menu/{id}', [UserController::class, 'getUserMenuItems']);
+            Route::post('/menu/{id}', [UserController::class, 'addUserMenuItem']);
+            Route::delete('/menu/{id}', [UserController::class, 'deleteUserMenuItem']);
+            Route::get('/submenu/{id}', [UserController::class, 'getUserMenuSubItems']);
+            Route::post('/submenu/{id}', [UserController::class, 'addUserMenuSubItems']);
+            Route::delete('/submenu/{id}', [UserController::class, 'deleteUserMenuSubItems']);
+            Route::patch('/password/{id}', [UserController::class, 'updatePassword']);
+        });
+
+        /**
+         * Route middleware menu user untuk akses delete
+         */
+        Route::group(['middleware' => ['access.user:delete']], function () {
+            Route::delete('/{id}', [UserController::class, 'destroy']);
+            Route::delete('/truncate-tokens', [UserController::class, 'truncateTokens']);
+        });
     });
 });
